Rename PodVisitor to avoid repetitive names
